Move example_peripheral and sensor_tag_config to a dedicated directory
