+++ example/lib/test_scenarios/test_scenarios.dart
@@ -6,7 +6,7 @@ import 'dart:typed_data';
 
 import 'package:flutter_ble_lib/flutter_ble_lib.dart';
 
-import '../sensor_tag_config.dart';
+import 'package:blemulator_example/example_peripherals/sensor_tag_config.dart';
 import 'package:flutter_ble_lib/internal/constants.dart';
 
 part 'base.dart';
